rewrite transfer data format
