+++ lib/interface/serializable.dart
@@ -1,4 +1,3 @@
 abstract class Serializable {
   String serialization();
-  Map<String, dynamic> serialization_json();
 }
