+++ lib/models/transfer_data/auth_data.dart
@@ -1,0 +1,20 @@
+import 'dart:convert';
+
+import 'package:invenza/interface/serializable.dart';
+
+class AuthData implements Serializable{
+  final String account;
+  final String password;
+  AuthData(this.account, this.password);
+
+  @override
+  String serialization() {
+    return jsonEncode(
+      {
+        'account' : account,
+        'password' : password,
+      }
+    );
+  }
+
+}
