+++ lib/models/transfer_data/forgot_password_data.dart
@@ -1,0 +1,18 @@
+import 'dart:convert';
+
+import 'package:invenza/interface/serializable.dart';
+
+class ForgotPasswordData implements Serializable {
+  final String email;
+
+  ForgotPasswordData(this.email);
+
+  @override
+  String serialization() {
+    return jsonEncode(
+      {
+        'email' : email,
+      }
+    );
+  }
+}
