+++ lib/models/transfer_data/issue_report_data.dart
@@ -1,0 +1,20 @@
+import 'dart:convert';
+
+import 'package:invenza/interface/serializable.dart';
+
+class IssueReportData implements Serializable {
+  final String issue;
+  final List log;
+
+  IssueReportData(this.issue, this.log);
+
+  @override
+  String serialization() {
+    return jsonEncode(
+      {
+        'issue' : issue,
+        'logs' : log,
+      }
+    );
+  }
+}
